requirements.txt
torch
torchvision
numpy
pillow
scikit-learn
tqdm
open_clip_torch
timm
huggingface_hub
pandas
csv
matplotlib